+++ src/pl/plajer/villagedefense3/user/User.java
@@ -26,10 +26,13 @@ import org.bukkit.scoreboard.ScoreboardManager;
 import pl.plajer.villagedefense3.Main;
 import pl.plajer.villagedefense3.arena.Arena;
 import pl.plajer.villagedefense3.arena.ArenaRegistry;
+import pl.plajer.villagedefense3.database.FileStats;
 import pl.plajer.villagedefense3.kits.kitapi.KitRegistry;
 import pl.plajer.villagedefense3.kits.kitapi.basekits.Kit;
+import pl.plajer.villagedefense3.villagedefenseapi.VillagePlayerStatisticChangeEvent;
 
 import java.util.HashMap;
+import java.util.Map;
 import java.util.UUID;
 
 /**
@@ -45,8 +48,8 @@ public class User {
     private boolean fakeDead = false;
     private boolean spectator = false;
     private Kit kit = KitRegistry.getDefaultKit();
-    private HashMap<String, Integer> ints = new HashMap<>();
-    private HashMap<String, Long> cooldowns = new HashMap<>();
+    private Map<String, Integer> ints = new HashMap<>();
+    private Map<String, Long> cooldowns = new HashMap<>();
 
     public User(UUID uuid) {
         scoreboard = scoreboardManager.getNewScoreboard();
@@ -116,10 +119,16 @@ public class User {
 
     public void setInt(String s, int i) {
         ints.put(s, i);
+
+        VillagePlayerStatisticChangeEvent villagePlayerStatisticIncreaseEvent = new VillagePlayerStatisticChangeEvent(getArena(), toPlayer(), FileStats.STATISTICS.get(s));
+        Bukkit.getPluginManager().callEvent(villagePlayerStatisticIncreaseEvent);
     }
 
     public void addInt(String s, int i) {
         ints.put(s, getInt(s) + i);
+
+        VillagePlayerStatisticChangeEvent villagePlayerStatisticIncreaseEvent = new VillagePlayerStatisticChangeEvent(getArena(), toPlayer(), FileStats.STATISTICS.get(s));
+        Bukkit.getPluginManager().callEvent(villagePlayerStatisticIncreaseEvent);
     }
 
     public void setCooldown(String s, int seconds) {